requirements.txt
streamlit==1.27.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
plotly==5.16.1
requests==2.31.0
pillow==10.0.0
